+++ tests/testthat/test-surveys.R
@@ -1,3 +1,5 @@
+context("surveys")
+
 test_that("surveys() gives expected output", {
   svy_tbl <- surveys()
   expect_named(svy_tbl, c("id", "name", "owner_id", "last_modified",